pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbenchmark"
version = "0.1.0"
description = "LLM Benchmark Tool"
authors = [{ name = "jamesnulliu", email = "jamesnulliu@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "torchaudio",
    "torchvision",
    "transformers>=4.42",
    "modelscope",
    "addict",
    "accelerate",
    "easydict",
    "ruamel.yaml",
    "pandas",
    "flash-attn",
]

[tool.setuptools]
packages = ["llmbenchmark"]
package-dir = { "" = "." }